+++ scripts/switch/push-docker-build-chiaki.sh
@@ -27,5 +27,4 @@ docker run --rm \
     -ti -p 28771:28771 \
     --entrypoint /opt/devkitpro/tools/bin/nxlink \
     docker.io/xlanor/chiaki-ng-switch-builder:latest \
-    "$@" -s /build/chiaki/build_switch/switch/chiaki-ng.nro
-
+    "$@" -s /build/chiaki/build_switch/switch/cloudpad.nro